Add and use an input parsing function
Introduce an input parsing function which accepts a stream, and call it
with either the input file or standard input as appropriate.  Store the
pattern in a member variable to avoid a return statement, consistent
with other functions and style.

On the command line, indicate standard input with the special filename
`-`, just as for writing to standard output.